+++ src/main/java/com/example/block/global/config/WebMvcConfig.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -35,6 +36,14 @@ public class WebMvcConfig implements WebMvcConfigurer {
         registry.addInterceptor(new UserIdInterceptor())
                 .addPathPatterns("/**")
                 .excludePathPatterns(Constants.NO_NEED_AUTH_URLS);
+    }
+        @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")  // 모든 경로에 대해
+                .allowedOrigins("http://localhost:3000")  // React 앱의 주소
+                .allowedMethods("GET", "POST", "PUT", "DELETE")  // 허용할 HTTP 메서드
+                .allowedHeaders("*")  // 모든 헤더 허용
+                .allowCredentials(true);  // 인증 정보 허용
     }
 
     @Bean
